Extract ConfigurationCacheJavaSerializationIntegrationTest
from ConfigurationCacheIntegrationTest
