+++ subprojects/configuration-cache/src/integTest/groovy/org/gradle/configurationcache/ConfigurationCacheIntegrationTest.groovy
@@ -439,92 +439,6 @@ class ConfigurationCacheIntegrationTest extends AbstractConfigurationCacheIntegr
         ImmutableMap  | "ImmutableMap.of(1, 'a', 2, 'b')" | "[1:a, 2:b]"
     }
 
-    def "restores task fields whose value is Serializable and has writeReplace method"() {
-        buildFile << """
-            class Placeholder implements Serializable {
-                String value
-
-                private Object readResolve() {
-                    return new OtherBean(prop: "[\$value]")
-                }
-            }
-
-            class OtherBean implements Serializable {
-                String prop
-
-                private Object writeReplace() {
-                    return new Placeholder(value: prop)
-                }
-            }
-
-            class SomeBean {
-                OtherBean value
-            }
-
-            class SomeTask extends DefaultTask {
-                private final SomeBean bean = new SomeBean()
-                private final OtherBean value
-
-                SomeTask() {
-                    value = new OtherBean(prop: 'a')
-                    bean.value = new OtherBean(prop: 'b')
-                }
-
-                @TaskAction
-                void run() {
-                    println "this.value = " + value.prop
-                    println "bean.value = " + bean.value.prop
-                }
-            }
-
-            task ok(type: SomeTask)
-        """
-
-        when:
-        configurationCacheRun "ok"
-        configurationCacheRun "ok"
-
-        then:
-        outputContains("this.value = [a]")
-        outputContains("bean.value = [b]")
-    }
-
-    def "restores task fields whose value is Serializable and has only a writeObject method"() {
-        buildFile << """
-            class SomeBean implements Serializable {
-                String value
-
-                private void writeObject(java.io.ObjectOutputStream oos) {
-                    value = "42"
-                    oos.defaultWriteObject()
-                }
-            }
-
-            class SomeTask extends DefaultTask {
-                private final SomeBean bean = new SomeBean()
-
-                @TaskAction
-                void run() {
-                    println "bean.value = " + bean.value
-                }
-            }
-
-            task ok(type: SomeTask)
-        """
-
-        when:
-        configurationCacheRun "ok"
-
-        then: "bean is serialized before task runs"
-        outputContains("bean.value = 42")
-
-        when:
-        configurationCacheRun "ok"
-
-        then:
-        outputContains("bean.value = 42")
-    }
-
     @Unroll
     def "restores task fields whose value is service of type #type"() {
         buildFile << """
